backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
supabase==2.3.7
python-dotenv==1.0.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
ruff==0.2.1
pytest==8.0.0
PyJWT==2.8.0
httpx==0.26.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.7